able to send contact requests
